♻️ G4EWJG guard: decompose commit implementation
