+++ .agentplane/tasks/202605282032-G4EWJG/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 G4EWJG task: Guard commit implementation decomposition [202605282032-G4EWJG]
