+++ .agentplane/tasks/202605282032-G4EWJG/pr/review.md
@@ -1,0 +1,39 @@
+# PR Review
+
+Created: 2026-05-28T20:32:58.005Z
+
+## Task
+
+- Task: `202605282032-G4EWJG`
+- Title: Guard commit implementation decomposition
+- Status: DOING
+- Branch: `task/202605282032-G4EWJG/guard-commit-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282032-G4EWJG/README.md`
+
+## Verification
+
+- State: ok
+- Note: Guard commit decomposition verified locally.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T20:32:58.005Z
+- Branch: task/202605282032-G4EWJG/guard-commit-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../src/commands/guard/impl/commit-close.ts        | 231 +++++++++++++
+ .../src/commands/guard/impl/commit-runner.ts       | 150 ++++++++
+ .../agentplane/src/commands/guard/impl/commit.ts   | 376 +--------------------
+ 3 files changed, 387 insertions(+), 370 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
